2/requirements.txt
numpy
tensorflow
opencv-python
matplotlib
pylab-sdk
tfa-nightly
Pillow
mtcnn
scikit-learn
